requirements.txt
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
Pygments==2.13.0
rich==12.6.0
shellingham==1.5.0
typer==0.7.0
openai==0.25.0
python-dotenv~=0.21.0